[roll] Roll fuchsia [netemul] Shutdown realm before continuing with test teardown
This test was prone to flakes because sometimes the dhcpv6-client would
still be in the middle of starting up when the test finished. Test
interfaces get removed when the fuchsia.netemul.network/Endpoint channel
to them is dropped, and this sometimes meant an interface was removed in
between dhcpv6-client creating a socket and trying to bind it to that
interface. This race led to unexpected error logs that caused test
failures.

In order to prevent realm shutdown from racing with test interface
teardown, block on realm shutdown before the channel handles go out of
scope.

Original-Fixed: 82913
Test: fx test netemul-v2-tests
Multiply: netstack-management-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576922
Original-Revision: d20b670738f4f161035ea827b66756694ecae5a2
GitOrigin-RevId: fe57e3fddc5fa25121d0e87c8626b4707031878c
Change-Id: I0a2979391d26a7622a5bdf80930a41ffbef83202