[roll] Roll fuchsia [audio] Fix conversion warnings
We forgot to file our own bug for this. The original bug was
fxbug.dev/58162, which is now closed.

These warnings revealed a bug in OutputProducer, which was not correctly
truncating SIGNED_24_IN_32 values. This bug has been fixed. Fixing this
bug requires adding an operation (either a shift-by-constant or
logical-and-by-constant). Some profiling with audio_mixer_profiler shows
that the impact of this change is basically negligible.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542986
Original-Revision: a9f7f00fa33ba1fc158f11b306f9254400db4460
GitOrigin-RevId: 6e0b021ddacf03a29e74b96bec55853a3a5b5c20
Change-Id: I4abfec685211fb125852f53bc60f6c0fc1e557bc