[roll] Roll fuchsia [camera][ge2d] Fix wrong exported format index
The wrong format index is exported at the output of the GE2D
stage of Sherlock configuration 0. The setting of the format
value by SetResolution will only set the input when the mode
is the overlay (in-place) mode. This means that the output
resolution is actually the input resolution, but since only
the input mode is set, the output format index is left as an
invalid value.

Please note that this invalid GE2D output format value is
never used in any Sherlock configurations so there is
technically no impact.

Original-Bug: 197900636
Test: Run normal unit tests to make sure there is no
      accidental side effect.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/578187
Original-Revision: e80c6f3943d7165d08581c091ba8bd05a683aebb
GitOrigin-RevId: af4840be5e3e3f28f04e0581d843d3159da30c89
Change-Id: I12a824e58917a0a516ef65e43554ce59b0a1b628