[roll] Roll fuchsia [testrunner] Set IO timeout for checking test start signature.
We have a context timeout that will time out if we are able to keep
reading from serial but can't find the start signature, but if there is
nothing to read from serial, it will hang until it reaches the io
timeout which is currently set to the test timeout. For checking the
start signature we should set a shorter timeout and reset it to the test
timeout once we've confirmed the test was started.

Test: testrunner times out and reruns the test when it can't see the
start signature after 5 seconds (changed the signature to look for to
force it to time out) https://chromium-swarm.appspot.com/task?id=55e60f5215ce6b10

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579755
Original-Revision: 7d32e76d8f2e72c92260e8ece654eab64e9b90a2
GitOrigin-RevId: e8ddeaa478e6992f09b03d735687a6adad8b9d5c
Change-Id: Ib92cf610022c9ebe8d3a63957e12f0121653aafc