[roll] Roll fuchsia [inspect][stats] Add counter for failed allocations
When we run out of memory and cannot add a new page with new blocks we
just return an error. In order to export how often this happens, we're
adding a counter that reports how many failed allocations we've had
since the heap was created.

Original-Bug: 75729
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579923
Original-Revision: 4420965a413d1447ee462fc1c55db75342e56803
GitOrigin-RevId: 1040e36f7c88a4eafdc1da447e4ad59fc54a245c
Change-Id: I639a4083c7d6a1e779994358d57d4ab04aa1cedc