[roll] Roll infra/recipes [fuchsia] Use vpython for running perfcompare.py
Use vpython for running perfcompare.py so that the .vpython file will
get used for specifying the versions of Python, numpy and scipy to
use.

I had thought that the recipes system always uses vpython, but that
seems not to be the case, so I don't know which versions of numpy and
scipy this was picking up before.

Original-Bug: none
Test: checked that vpython is used for the "summary of perf test
  results" step on the CQ's terminal.x64-release build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/579446
Original-Revision: c87101d25ec064aab9002ee192bfb3876a9bffe2
GitOrigin-RevId: 7fa24af4e3409d72632618463d852b4fee4c7d4f
Change-Id: Ie65f9656cbb72301d590531545690b15e1cc6988