[roll] Roll fuchsia [io] Require fuchsia.io to have explicit oridinals
Previously, the FIDL compiler had some magic to rename fuchsia.io
ordinals to fuchsia.io1. This magic was intended to make it easier to
migrate to fuchsia.io2 by letting the new io2 versions of the methods
have the "normal" ordinal.

Now that we are in the midst of the transition, we don't want to be
confused by this magic. This CL requires that all ordinals in fuchsia.io
be written explicitly in the protocol definition, removing the magic. To
ensure that we do not miss anything, the compiler now crashes whenever
it would have previously applied the magic.

Original-Bug: 77623
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581401
Original-Revision: 2a953a2242cf5ff5a8e3a6d363c17aee19b6422e
GitOrigin-RevId: c99fceeacd5894794feef0a054f6c594628e0232
Change-Id: I0d47d335b7b53a5dc990c7852037fdd2ca08b42f