[roll] Roll fuchsia [netstack] Fix deadlock while adding address
`ifState.mu` must be locked before `addressStateProviderCollection.mu`,
but `Control.AddAddress` acquires the latter lock before adding an
address, and after adding an address the logic to send an interface
property change event is triggered, which acquires the former lock.

This has been fixed by not acquiring `ifState.mu` when enqueuing the
event as a result of adding an address when serving
`fuchsia.net.interfaces/Watcher`. Care must be taken to only do so for
IPv4 addresses and IPv6 addresses when no DAD is performed, as an IPv6
address for which DAD is performed should only be visible to clients
when it is no longer tentative.

While I'm here, refactored some of the functions defined on `Netstack`
by introducing functions on `interfaceWatcherCollection`.

Test: fx test netstack-gotests \
              netstack-fidl-integration-test \
              netstack-ipv6-integration-test
Multiply: netstack-fidl-integration-test: 30
Original-Fixed: 84310
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579841
Original-Revision: 0b87eaba9aded381c772eb05e731b98167438a2c
GitOrigin-RevId: eb89ca530aa8b37a3a63fb5fae76cdf65a5c2e59
Change-Id: Ibd90aa49b9e351fd471fca264cfa447e0bd34a94