[roll] Roll fuchsia [ffx][log] Fix disconnect/reconnect loops.
The strategy here is that upon receiving a disconnect from the backend,
we first check to see if there is a log entry present following the
disconnect event. If there is, we just output the disconnect event and
keep going without triggering a retry loop. This prevents cycles of
disconnect/reconnect triggered by a single disconnection.

Original-Fixed: 73714
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581343
Original-Revision: 548c12cfbd0fcc2b307f33bc3346222106300473
GitOrigin-RevId: cc7541ff6ee601923415d19b9d649edef1c8573f
Change-Id: I9d9c0e1c8c8d104db1d8b1cbd48027e587b33205