[roll] Roll fuchsia [wayland] Flatland scale and crop.
Refactor image import to support multiple surfaces
and flatland instances. This also enables crop and
scale support as that's trivial when we can create
unique image instances for each surface.

Note: A bug in Sommelier is preventing cropping from
working as expected. The source rectangle is not
correctly adjusted for the buffer scale. weston-scaler
can be used to verify that this works as expected when
that bug has been fixed.

Original-Bug: 78911
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581562
Original-Revision: 08f92c58430e3466e9d7e255eb1bd9173f2aa72b
GitOrigin-RevId: 07c8824c5acb546fbf59b67f40fd0606d28f17b0
Change-Id: I9c81ba006bc8b8c2e36f7a383be43c10606ea038