[roll] Roll fuchsia [aml-pwl-regulator] Copy fields of FIDL table to avoid lifetime bug
The existing AmlPwmRegulator is constructed by moving a FIDL table
object into the AmlPwmRegulator. However, this move only moves the
top-level table header and not the contents of the table, which
only lives until the fidl::DecodedMessage object goes out of
scope. Therefore, accesses to the vreg_range_ field after Create()
exits are use-after-free.

This CL fixes this issue by copying the contents of the table
field-by-field to the AmlPwmRegulator object.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581701
Original-Revision: 6f79773b84b28d91ea080bfea29127e0a9643787
GitOrigin-RevId: c428717947d8753538d7092b2450c1e8ca2af62d
Change-Id: Ia7f3a6b2ee4b1b36d19d8570a518c2dcdc9a2d57