[roll] Roll fuchsia [fatfs] Fixed a port overflow
The implementation of mark_dirty was causing too many tasks to be
spawned which could end up overflowing the port. This CL changes the
implementation so that it doesn't spawn any new tasks whilst an existing
task is pending, but it comes at the expense of spuriously waking up the
task every second (which should be fine).

Original-Fixed: 75606
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581682
Original-Revision: 19886740df06fc8464713e89a85c6c901261ed6e
GitOrigin-RevId: cef26c352935cbb032cd4d56599824b21bb4cd57
Change-Id: Ibea5e090d8e18e1b287833cb506f799802678b64