[roll] Roll fuchsia [storage] Fixed racy link
The handling of link in the Rust VFS was racy. This CL adds a test and
fixes it.

Avoiding the races and keeping it working for pseudo directories is hard
because the source can be a number of different concrete types and locks
need to be maintained.  Support for hard links in mutable
pseudo-directories is not actually needed right now, so I've removed
support for that for now.  If the need arises, we can add back support
but it will involve a fair amount of restructuring.

Original-Fixed: 81237
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581182
Original-Revision: a6e270822b1f887d0a10ab630de4a99fbd6ccaf5
GitOrigin-RevId: 65a69c076cafda6e9a5dfef426199d140ee58745
Change-Id: I3cbd5605a347791b4d85516f9054f12e0731cdd3