[roll] Roll fuchsia [netstack3] Move socket maps to sockets module
Move `ConnAddrMap` and `ListenerAddrMap` from the `transport` module to
the `sockets` module, and rename them to `ConnSocketMap` and
`ListenerSocketMap`. `ConnSocketMap` is used in ICMP, so it doesn't make
sense for it to live in the `transport` module.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581761
Original-Revision: e43d7c4aa1fa2e5b62ae1c137e5a141dc441f38f
GitOrigin-RevId: 6e877c17fc4756cddc0aa6b0cadf8c6b524da86a
Change-Id: Ie38601fe4494e320fc89dca9d7b2565fe17dd8ec