[roll] Roll fuchsia [netsvc] Temporarily disable netsvc stress tests on ASAN builder
FIDL is undergoing a wire format migration. As part of this, a
transformer is used to transform between the old and new wire
format before/after decode/encode. This slows down the netsvc
stress test so that it keeps timing out.

This CL disables the test on ASAN builds until the migration is
complete.

Original-Bug: 84697
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581902
Original-Revision: 8199b540257dc361e52ade4f5d6f0db48768923f
GitOrigin-RevId: 426b76684f228f90268f4463cd675775a27ad761
Change-Id: Id8d69e708c6785325a179dc73af98dc4f94f1c81