[roll] Roll fuchsia [bt][common] Don't assert on too many AD UUIDs
AdvertisingData is represented over the wire in TLV format, so there is
an inherent size limit for each field value based on the maximum value
of the uint8 length field (i.e.uint8_t::MAX). This means that for each
instance of a "per-compact-size service UUID" type field as defined in
BT CSS v9 Part A 1.1.1, there is a bound on the number of UUIDs that fit
in that field. An assertion introduced in fxrev.dev/579904 incorrectly
assumed that peers only ever send us one "per-compact-size service UUID"
type field per AD. ibid. says that is the case for valid ADs, but AD is
untrusted peer data, so there is no guarantee this is true. See fuzzer
bug for further information.

This is morally a revert of fxrev.dev/579904, but updated based on
fxrev.dev/579905 along with an extra test+comments.

Original-Fixed: 84487
Test: Modified bt-host-common-tests with AdvertisingDataTest.
      TooManyUuidsOfSizeRejected. Verified that fuzzer test case no
      longer reproduced with this patchset.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582102
Original-Revision: 7d4ad08ba7481470edf646f8b8dcabfb2ffdd4aa
GitOrigin-RevId: b4a7d82f3fff2739535a9fd416dacb744fe382bb
Change-Id: Iea5cd6a39a74e3fb7fcb81b8acc7984314bb9878