[roll] Roll fuchsia [kernel] Send MP_IPI_INTERRUPT when Killing or Suspending threads
When killing or suspending a thread, issuing an MP_IPI_RESCHEDULE is
unnecessary.  We don't need to cause the target CPU to call into the
scheduler.  We only need it to take an interrupt because the interrupt
handler will check and process thread signals as needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581381
Original-Revision: 7e70d2ba2900cc90f6ea64bd43247afdcaa6ea32
GitOrigin-RevId: f207e62990ec517727d3a1d46c3cfec95b5bc1bb
Change-Id: Ib1fb043389c87307b54717f718fef04d2d4ba195