[roll] Roll fuchsia [llcpp] Define message encoding / decoding classes outside of type class
The class definitions of structs, tables, requests and responses are
made busier than necessary because they contain the definitions of
UnownedEncodedMessage, OwnedEncodedMessage, DecodedMessage within them.
This makes it harder to find and read the core parts of the type
definition.

This CL moves these definitions outside of the body in order to make the
classes easier to read.

Original-Fixed: 84425
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582262
Original-Revision: 36c35169a2d812f10113bfe8ad3930e07b8970dc
GitOrigin-RevId: e411a2cb700816d460b4ffd3f349f24137a18425
Change-Id: Ic11d6a63c792ce9a031602ea71a393e6f88f6c05