[roll] Roll infra/recipes [archive] Fix target paths in archive tree to be in a consistent format.
Paths like Path('path', 'to', 'file') and Path('path', 'to/file') are
counted as separate paths in the link_map in the archive tree. However
when create_links() gets called, it converts the paths in the map to
strings and different entries with targets that have the same string
path but different Path end up overwriting each other instead of
appending the linknames for the same target. This change makes the paths
we provide as target paths into a consistent format so that different
variations of the same string path will all point to the same path.

This fix is necessary for using ffx in infra because we'd add ffx to the
task inputs as a tool in the root directory of the tree as well as a
shard dep for some ffx tests.

The linkname cannot be converted in the same way because there is a
check in SymlinkTree for tree.root.is_parent_of(linkname) which will
break if linkname goes through abs_to_path() because abs_to_path() puts
the entire path into one piece of the Path and it seems tree.root needs
to be a separate piece of the Path.

Original-Bug: 84714
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/582361
Original-Revision: 3d0f5f391f70b32bf58cb9850187ee8fcab56b6e
GitOrigin-RevId: 95cf59298d44643b99a90e3c6ee97e5d881e39e0
Change-Id: If31c5b0415f75910549961365f776d4ac9c7531d