[roll] Roll fuchsia [net-types] Prefer Witness::get over into_addr
For `Witness<A>` types, when both `get` and `into_addr` are available
(which is the case when `A: Copy`), prefer `get`. Document this
preference in the docs for `Witness` and update call sites to respect
this preference.

Also remove "into" methods on other types where possible.

Original-Bug: 83331

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573843
Original-Revision: 94a043cb48f34ddc0ec5d7beb8949ea3e4ea2b24
GitOrigin-RevId: b579dc1b3d3af118fad7f66f9d834711a7a4b00f
Change-Id: I789ef5d3f176e2e9dea32ac225af74739024b40e