[roll] Roll fuchsia [f2fs] Add range checkers for checkpoint packs
This CL adds checking the validity for some fields that
a checkpoint pack have such as the sizes of nat/sit version bitmaps
and the next free node id. If they are invalid, it stops mounting
since it means that there is corruption on the read-only superblock
or the checkpoint pack.

Original-Fixed: 84609, 84611

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582621
Original-Revision: 7f72538230dee30296ffe14d12edb1fe39c012fa
GitOrigin-RevId: 6097c4b82bbd563bfaa4802fc2f2daa752148f3c
Change-Id: I2da14d8619673390e1ffa9968991dad97bef7807