[roll] Roll fuchsia [fidlc] Sometimes we have no ordinals on a method
Due to previous compilation errors, we might not have an ordinal on a
method, and must therefore protect the access.

Implementing the restriction on `fuchsia.io` requiring explicit ordinals
as a compiler check, rather than an assert. This avoids the compiler
failing in a more obscure way, and is more in line with how fidlc is
meant to operate.

Original-Bug: 77623
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582341
Original-Revision: b7866d4c61e705d49589b6b06556ea011a3789d8
GitOrigin-RevId: c46d724d44ff8be9006f93fbc9cc49a376227407
Change-Id: I84da4b3e904770522b7e01f12590686df84cdc55