[roll] Roll third_party/go [go] Increase go_os_test timeout to 3m
FIDL is currently undergoing a wire format migration. As part
of this, a transformer is used between the old and new wire
format which adds extra latency. This has the effect of slowing
down the go_os_test on the ASAN build just enough that it exceeds
the existing 2 minute timeout.

This unblocks:
https://fuchsia-review.googlesource.com/c/fuchsia/+/574261

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/581903
Original-Revision: c5f39976d1991f7ca31f5712f489c5fb9b5bbc89
GitOrigin-RevId: 9a5b370d4320343f01d9a44366dc74b3fb1909e3
Change-Id: I6040b8f440c9ce86d586b9ecb22b0328f60dd91e