[roll] Roll fuchsia [appmgr] Temporary fix for cache purging
Appmgr's notion of free space is not correctly computed for FVM-enabled
filesystems, and doesn't take into account the budgeted maximum size of
the partition, so cache purging might never kick in (since the
denominator is higher than expected, and we purge when using 95% or more
of the partition).

This CL is a temporary fix for this issue which addresses the problem
with minimal churn.  A proper fix is impending but requires plumbing
much more information through the system, so this fix will take up the
slack in the meantime.

This fix is only enabled if the new GN argument
`minfs_board_maximum_data_bytes` is set.  This CL does not set this GN
argument for any devices.

Original-Bug: 83978
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582190
Original-Revision: e1d1449be712540fb0e1f0a4c0dca4dd658318d7
GitOrigin-RevId: 5ce159d85dfe324fe691ad7ce7f7db40f7ef613f
Change-Id: I22cbd922fc8849720e09a71db071818ef685b378