[roll] Roll fuchsia [fidl][llcpp] Completely remove WaitForChannel
The feature to block the current thread until all sync calls on a
`fidl::WireClient` complete and extract the channel had been removed a
while ago due to deadlock risks and lack of use. This CL completely
removes this concept from the binding internals, resulting in simpler
code and easier to reasonable ownership. Now a `ClientBase` weakly
references a `AsyncClientBinding`, which has a shared `zx::channel` to
which it may vend out new shared references to support in-progress sync
calls.

Under this ownership model, the channel will be destroyed as soon as the
binding tears down (provided there are no sync calls), regardless if the
user destroys their client objects. This would help reduce resource
leaks.

The current semantics is that if the user destroys or teardowns a
`Wire{Shared}Client` while there are ongoing sync calls, those sync
calls would be allowed to finish, after which the channel is finally
destroyed. In the future we would like to explore eagerly interrupting
those sync calls in this case (fxbug.dev/78906).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582043
Original-Revision: 0d04a4bc93d29c028e2647ee7236af350df1462f
GitOrigin-RevId: 80261d4571054d699f82b26f173a8e8a4ad4139f
Change-Id: Ifd4d5acf07eb67d6453a1ed464c49af547d6d321