[roll] Roll fuchsia [settings] Migrate to inspect logger
Now that we have the capability to log via the inspect logger class,
the remaining usages of the inspect agent can be migrated to use the
inspect logger. This changes the remaining usages of the inspect
agent to use the inspect logger.

Original-Bug: 84416
Test: fx test setui_service_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581514
Original-Revision: 4e20da3687b1e292a95a0b75a4c7143cd342caa9
GitOrigin-RevId: b2ee05f3b18caa746a586bc6faa4b94c7a337dd4
Change-Id: I094d36a371eb8636e450e4a2455672b46de5d5cf