[roll] Roll fuchsia [cts] Update verify_release script
Now written in python and moved to
//sdk/cts/build/scripts/verify_release.py.

This can be run as a test or standalone as a script, though this likely
will never be able to execute as-is in CQ since it runs fx commands.

Following CLs will add a step to run the tests at the end, and reset
your "fx args" settings to be less disruptive.

Original-Bug: 84049
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573941
Original-Revision: 9467b7079fa42957103a973013d8b98c1e31861a
GitOrigin-RevId: cae267d892a40487c99accca3591e1ef5cb0d67f
Change-Id: I726b8b1889efa53f1a6a12792609c12e0a9427d9