[roll] Roll fuchsia [virtualization][vsock] Avoid sending spurious shutdowns to endpoints
virtio-vsock responds to unexpected incoming packets by replying with
a RESET response. To implement this, the receive path allocates a new
NullConnection and places it in the SHUTDOWN state. When the transmit
path of virtio-vsock processes this NullConnection, it will send out
a RESET packet and destroy the NullConnection.

However, when the transmit path destroys this connection, it also
notifies any connected endpoint that a socket has been destroyed. The
guest manager reacts to this event by freeing the port associated with
the connection.

In the case of an unexpected packet being received, however, no port was
actually ever allocated: this leads to the guest manager freeing a port
that may be in use by an unrelated socket.

Update virtio-vsock to avoid notifying endpoints and closed
NullConnections.

Original-Bug: fxbug.dev/84286

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582625
Original-Revision: 8e9ea7a72b006877a55419876ee429ef0e554d6f
GitOrigin-RevId: 94e5367c7227d54fe5c5b6a22f80b61e2866b6e2
Change-Id: I2d7f289696aeeb68d00c9c849a44aae7f697c114