[roll] Roll fuchsia [fvm] Tidy up the FVM tool a bit
Fixed some quirks with the command-line parsing (some would cause the
tool to crash).

In the process of doing this, I discovered that the ffx was invoking the
fvm tool incorrectly (to be fair, the fault lies with the fvm tool which
is a bit clunky), and it was passing arguments that would be ignored.

Original-Bug: 62215
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580603
Original-Revision: adb83e9642827900aee94863f03c57be65a3e3b4
GitOrigin-RevId: 260a7fcc5f1a934dda3f515ac16313f0a0691563
Change-Id: I3b8e577cd9b4184bb42bcd6b951eec7d4561e0f7