[roll] Roll fuchsia [starnix] Prevent "rm foo/.." and "rmdir foo/.."
Previously, we would look up the child twice in the unlink codepath,
once in NamespaceNode and again in the DirEntry code. We need to perform
the lookup exactly once in case another thread is modifying the file
system concurrently. This CL centralizes the lookup in the DirEntry
code.

In order to handle mount points correctly, we need to plumb the
NamespaceNode for the parent to the DirEntry code. This CL also follows
the same pattern for rename.

This code is still slightly incorrect because we do not hold the lock on
the mount point table for the duration of these functions. Another
thread could potentially make the directory entry into a mount point
while we are trying to unlink or rename the entry. This CL includes a
TOOD for that issue, which will need to be resolved in a future CL.

Original-Bug: 79440
Test: CannotRemoveDots
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582851
Original-Revision: 2f1626db258b07a62cf119c20c0e4963f85aa0a3
GitOrigin-RevId: 37e0f0d46212c16256adb26387b790c95a55b480
Change-Id: I30069ee3aff91263658f4354650f1b4d7cfddc17