[roll] Roll fuchsia [fidlc] Add type_shape_v* to type field of JSON IR
There isn't currently a universal way to get type sizing information for
types in the JSON IR. Bindings may redo the logic by combining size
information on specific decl types with special logic handling arrays,
vectors, etc.

In particular, this is useful for implementing different logic on
members of unions and tables based on the type size. Currently, it is
necessary to reimplement some of the type shape logic to get this
information. When type shapese are in the IR under each type, it is much
more straightforward to extract it.

Original-Bug: 79578
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582663
Original-Revision: ba0f50bbcbf658c0f63c3062a1f0f13d447f7b7d
GitOrigin-RevId: a08cc5b3fa01484d9a8708d421769313411dcbbd
Change-Id: I1a82be92d3aaf74d1a99eaaa351c076400a7c2a7