[roll] Roll fuchsia [fshost] Add support for binding nand-broker to nand devices.
This CL extends the block watcher to also watch for NAND devices.
On Chromebook, a userspace component (//src/storage/flashmap) is
responsible for interacting with the NAND, and it uses the nand-broker
to do that. This CL eliminates the need for its clients to handle
binding the nand-broker manually.

Multiply: fshost-integration-test
Multiply: paver-test
Original-Bug: 81685
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582121
Original-Revision: 74372ab6ca9576787e5613f56bea5846fc5ae4c6
GitOrigin-RevId: b9bcf4ce24842edfe7daa258a742e43cc588f7d6
Change-Id: Ibc39e4e220bbc9d47aa2057c78d2060fbfbb86a0