[roll] Roll fuchsia [sl4f] Add modular_facade.IsBasemgrRunning
This method provides a way to check if basemgr has started, either
as a session or legacy basemgr component.

Clients should use this facade instead of checking that the basemgr.cmx
component is running, as that is not compatible with sessions.

Original-Bug: 82392
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583644
Original-Revision: cf691cf1f0f2c27108a7c6326149c2b5ee20917c
GitOrigin-RevId: f50fec7bdd7aee10acc03ccfbfaff3f52ba6a350
Change-Id: I94c961d5924f209c1df9c9cdba9255bbe071e4d0