[roll] Roll fuchsia [kernel][crypto]: Add EntropyPool.
EntropyPool represents a collection of entropy from user defined
sources. Its purpose is to represent an opaque container for the raw
types used internally in the PRNG impl, and will be used to seed the
PRNG with 256 bits of entropy.

EntropyPool is a movable and copyable  object, that scrapes itself once its been
moved through mandatory_memset. The copyable is required for snapshoting
on concurrent contexts for calls to Add.

By doing so, we can guarantee that the Prng is always initialized with
enough entropy bits, and we can clean it up(remove the non-thread-safe
thread safe bits of its interface.

Test: k ut crypto-entropy-pool
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562149
Original-Revision: 5d717acec0b762558b4979f389196de3fc158de8
GitOrigin-RevId: 7e84da849f66b6f0a122acd6c5f988f9c1fdf61c
Change-Id: I67d7f7b83abf7f123b997328eb237f0d06835d54