[roll] Roll third_party/go [go] Increase go_os_test timeout to 5m
FIDL is currently undergoing a wire format migration. As part
of this, a transformer is used between the old and new wire
format which adds extra latency. This has the effect of slowing
down the go_os_test on the ASAN build enough that it exceeds
the existing timeout.

A previous CL (I7fa19fbe348492617abd796a328cb7a9f38e0415)
increased the timeout to 3m, but there is a long tail
of runs which take significantly longer than 3m. CQ was run 5
times in a row with a 5m timeout, with the test passing each
time.

This unblocks:
https://fuchsia-review.googlesource.com/c/fuchsia/+/574261

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/583205
Original-Revision: 61f22f2111ffab0317667a97c1d2e2f55cdbfe1b
GitOrigin-RevId: 1a16e832205f81257b2a80522f7fc8110a2ca015
Change-Id: I57ebeeb8415e78211edf68241f994d60ca4b2904