[roll] Roll fuchsia [cm_fidl_analyzer] Don't support dynamic offers.
Updating a TODO added in a previous CL which suggested that
cm_fidl_analyzer will need to support dynamic offers. In fact,
cm_fidl_analyzer is always running in a static context, so it's fine
to panic if it encounters a dynamic offer. This CL removes the TODO
and gives the panic! a clearer message.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582191
Original-Revision: 27cde5ac93b2a5fa091754e4db1dcb904ecec2c6
GitOrigin-RevId: 0cabd4417924650e49bac73c6d9a1e04d97ad2d7
Change-Id: Ia8bf7f5e69ec05b38c46741c7c75098012afd22a