[roll] Roll fuchsia [ftl] Remove inspect test and return integration test to CQ
Removes a test that broke as a result of Inspect behaviour. The
integration package was disabled from CQ as a result. This test mostly
just verified that the inspect nodes it creates at start up get created,
which is mostly a test of the Inspect library. So this test has been
removed and the package is being returned to CQ.

Since this was disabled we've had log severity restrictions added, so
this is disabled for this test as well since we actually want failure
to locate an NDM as an error log, even though it is normal for a new
device.

MULTIPLY: ftl-integration (fuchsia):20

Original-Bug: 84931
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583504
Original-Revision: 6595bb6252450be96874880d1494cd5214bef354
GitOrigin-RevId: 0a291e6a2ff4283f8ad78809841de8be1cce9199
Change-Id: Ia7f1e8e3592c41aaea358482bd5ae9c26fb15c44