[roll] Roll fuchsia [kernel][ktrace] Refactor for testing.
Refactor the way that things like syscall names, probe names, thread
names, and so on, get added to the trace buffer immediately following
a rewind or start operation to facilitate testing.  Specifically; move
these operations into protected virtual methods which can be
overridden by a test version of the KTraceState class, so that the
result of the operations does not depend on the current state of the
system.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579744
Original-Revision: ffd04c86827cd95e23efe86615c8748fae79c3ef
GitOrigin-RevId: d6e1ff0daa9f17f3851e87cde615796d1fb795b9
Change-Id: I61e151a3a7c9a11ac8ef264fc5084cd67eec1747