[roll] Roll fuchsia [fx format-code] Fail if unable to choose formatter
If a file doesn't have an extension or it has an unsupported extension,
`fx format-code` currently exits successfully without doing anything.
That's confusing because it makes it seem like the file is already
correctly formatted, when in fact fx doesn't actually know how to format
the file. A more user-friendly approach is to fail noisily in this
situation so the user knows that formatting failed, and why it failed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584061
Original-Revision: 2f162f63d172d7e950c669b8c73ef7472ae198a6
GitOrigin-RevId: ed4f980a110553fdc5942773294d608a878f5e80
Change-Id: Iedfcedf4de4ec055038af84632e4d940ea480a31