[roll] Roll fuchsia [kernel][crypto]: Rename PRNG->Prng.
All caps are usually reserved for macros. So its clear it is class for a
reader, when we follow this sort of pattern.
As a side not, GlobalPRNG is a namespace, so its been renamed to
global_prng to keep the same behavior. This is an uncommon pattern,
usually singletons are provided in the form of Class::Get/GetInstance.

Test: CQ, k ut all
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562762
Original-Revision: 721e88dc542519a762088d9bc15ddd0ab0dd8ebf
GitOrigin-RevId: 2a1ce1463b013d2b5d824eeb4a2e23888e9f0303
Change-Id: I07f34964d196c6ddc5307ac719ebf1bbb0b70473