[roll] Roll infra/recipes [code_search] Skip push in dry run mode
`git push --dry-run` will fail if a code_search tryjob races with a
production code_search build pushing the same commit content, so we
should skip pushing entirely in dry run mode.

This is indicative of a larger issue with this builder: the tryjob
commits and pushes to the same repo as the production builder so the
tryjob's behavior is nondeterministic and may vary with respect to the
state of the production builder. We worked around this issue for rollers
by setting up a static read-only repository for roller tryjobs to
operate on, but it's probably not worth doing that for the code_search
recipe since it's not quite as critical a builder, and it's likely to go
away relatively soon anyway when we switch from Jiri to submodules.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/584062
Original-Revision: 72fb72aae15f8bc8f97c176dc061174a8d0dc6c9
GitOrigin-RevId: e91385c18f1fa997ae15c7a59aae4d7ab060bf0e
Change-Id: I0d8d16e98f77d2347f15232debc4dd675309b003