[roll] Roll fuchsia [cts][session] Remove cts_session_config
The cts_session_config rule globally configures the startup session
for all instances of session_manager in a build. This will cause
session_manager to crash on startup if it can't resolve and start
the session.

For example, the session_manager in core will crash if
reference_session_config is included in the build because it references
a relative component that doesn't exist in session_manager's package.

In the future, tests should either instruct session_manager to start
a session with the `fuchsia.session` APIs or explicitly pass
configuration to it, and not rely on global configuration.

Original-Bug: 67088
Original-Bug: 83905
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584244
Original-Revision: 7bd7f8301876f5bd4d686790578a48063490f2da
GitOrigin-RevId: 423c24264a146ba2db059cc3789e920624a0f67d
Change-Id: Iacdb9d1e9e3924a32006ca72bb9bd927760c8cf5