[roll] Roll fuchsia [kernel][vm] Update access information for inactive pages
When pages in PageQueuePagerBackedInactive are accessed, their stashed
queue number in the vm_page_t is updated, but since this queue is not
processed as part of ProcessLruQueue, the pages are never moved out of
the inactive queue. This means that we might need to fix up the queue
when peeking pages for eviction. Without this fix we might have been
evicting pages that were in the inactive queue at some point but had
since been accessed.

Fixes an existing bug in eviction from the inactive queue, not directly
related to hinting, but required for hinting to work as intended.

Test: kernel-unittest

Original-Bug: 65334
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584163
Original-Revision: 9c9ca60684fc9cd4ba1d3eca5444ef95ad2d80b1
GitOrigin-RevId: 53acd53f683d0a1ed6a0059c015acc0538c2d71e
Change-Id: Id7bfe8f92105d56beab49b5b9e42aa6cc7f6a125