[roll] Roll fuchsia [llcpp] Switch to using v2 as the in-memory LLCPP format
This CL changes LLCPP envelopes (see envelope.h) to use v2 as the in
memory wire format. For the sake of reducing the size of the CL:
- The transformer is used to continue to encode v1 bytes on the wire
- The LLCPP API is unchanged

The idea is to change the in-memory format now and then gradually change
each of the pieces that interacts with it in the future.

Original-Bug: 79578
Original-Bug: 84116

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574261
Original-Revision: d99b1db6ddfb61dcd19964b47888e1169b793c77
GitOrigin-RevId: 1aebe424aabd0559e6d67e94962a5c6dd2e1c146
Change-Id: If1174be418c7401b74d3aa82ff774122ab9ba64b