[roll] Roll fuchsia [testrunner] Fix nil tester check.
In go, when checking if an interface is nil, it could either be the nil
interface or a non-nil interface with a nil value. Checking tester==nil
checks the first case, but when we set `tester, err =
newFuchsiaSSHTester()`, even if newFuchsiaSSHTester() returns nil as the
tester, tester gets set to a non-nil interface with a nil value. Thus
the tester==nil check doesn't work and we need to check the underlying
value. We can fix this by having the newFuchsiaSSHTester(),
newFuchsiaSerialTester(), and newSubprocessTester() funcs return the
tester interface instead of a pointer to their respective testers. That
way when we return nil from those functions, it doesn't set the pointer
value to nil but properly sets the tester interface value to nil so the
`tester==nil` check works.

Original-Bug: 85051
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584145
Original-Revision: d7565b2319c7549aa2a8984bf9071755f96098fc
GitOrigin-RevId: a3331392111d5619b02d183f809e7d7719e61579
Change-Id: Ia49d3f6605134f4ed5bc3d10426b0289c83235e4