[roll] Roll fuchsia [kernel][object] Remove resource usage from MsiAllocation
MsiAllocations are currently creating resource objects to track the
blocks of vectors they allocate. However, this isn't actually suitable
due to the difference between GSIs and Vectors on x64, or with the idea
of what a vector actually is. In this case, a vector is managed by the
interrupt controller in question (GIC or IoApic) and the resource is not
used for any sort of ownership exclusivity or transfer. With this in
mind, the platform level defines what an Interrupt Resource corresponds
to and MSI will leave the management of those objects to the platform
rather than risking conflating the vector space with something else.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575777
Original-Revision: b55acb77ebac6128fa375218c3fa5c52ca0c703b
GitOrigin-RevId: 7761c7f3bd011e5c7d429b14054e975b5a3c253f
Change-Id: Ic08d92b387796347fb060828a3946b64b9867cf1