[roll] Roll fuchsia [x64][interrupts] Configure available IRQs based on MADT tables
This adds a method `apic_io_get_gsi_range` to obtain the min
and max values for global system interrupts provided to us by
ACPI. Using this, we can properly initialize the ResourceDispatcher
corresponding to IRQs with the correct range of interrupts available
to the system. This fixes an issue where systems still in PIC mode
were using low IRQ #s that were outside the previously incorrect
IRQ range of [16, 255].

This is a re-land of fxr/569222 to return to patchset 14, rather than
the auto-submitted experimental (and wrong) patchset 15.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575778
Original-Revision: 3df93341119dbc0a21e2369535c84fb93670c329
GitOrigin-RevId: 70b79ecdc15c3ec859106a6c62fc987b706f53c1
Change-Id: I19f1d804fba1493d346021c50f46a887bc90ca86