[roll] Roll fuchsia [fx] Fix format-code warning logic
fxrev.dev/584061 made `fx format-code` fail and print an error if it
encountered a file it didn't know how to format.

This is useful for notification purposes if the user specifies a single
file that `fx format-code` can't format, but it's bad for `fx
format-code --git`, which tries to format all changed files. In that
case it's better to silently skip files that we don't know how to
format.

Likewise, the `--files` flag can take a glob, so it's possible that an
unformattable file is only included incidentally by a glob. Therefore,
we only show the error if the user specified a single file (or a glob
that only matched one file).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584344
Original-Revision: ed796ef621dc7523acbc606354dbeeaa965687d6
GitOrigin-RevId: 05c8c0bb1dde118968cdecb4b312edfc24950792
Change-Id: Ic7f9235177c2132496377065342359d12aef8d66