[roll] Roll fuchsia [cm_fidl_analyzer][routing] De/serialize AnalyzerModelError
In order to plumb errors from the routing static analyzer
through to Scrutiny, we'll need to be able to de/serialize
AnalyzerModelError and the various error types that it wraps.
The main intermediate step is to derive serde::{Deserialize,
Serialize} for RoutingError on host builds, replacing
'static borrows of error params with copies in a few places
to simplify lifetime management.

Original-Bug: 61861

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584603
Original-Revision: 998695acd63aa699ae118cfa0c8169cc37fcc1a4
GitOrigin-RevId: cf4da31efcc0435bb3b6c9d85bb85a47a0004598
Change-Id: Icaa8280f6d6bda9561745f305a948180def7817f