[roll] Roll fuchsia [audio][sherlock] Fix bitmask ordering
mpuryear@ found that we are incorrectly setting the bitmask used by
the SetActiveChannels API (used by the audio-driver-ctl -a parameter).
We invert ring_buffer_channels_to_use_bitmask such that:
The least significant bit corresponds to channel index 0
This per RingBuffer::SetActiveChannels() API specification, such that
the bitmask sequence matches the channel ordering in the data
ring-buffer itself.

This is not a change in the driver itself, but rather how we
configure the mapping in the board driver. Also we only change
the bitmask that is used for SetActiveChannel(), others are not
reversed in this CL.

The "/tmp/n" files below only have audio in the n channel.

Original-Fixed: 85229

Test: audio-driver-ctl play /tmp/2 plays in Sherlock's woofer
Test: audio-driver-ctl play /tmp/3 plays in Sherlock's left tweeter
Test: audio-driver-ctl play /tmp/4 plays in Sherlock's right tweeter
Test: audio-driver-ctl -c 4 -a 2 plays Sherlock's woofer
Test: audio-driver-ctl -c 4 -a 4 plays Sherlock's left tweeter
Test: audio-drvier-ctl -c 4 -a 8 plays Sherlock's right tweeter
Test: All speakers play audio in normal usage.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584681
Original-Revision: 3b91b36bb34a1cfd81216ecb444e435fe8598123
GitOrigin-RevId: bd4c5c3cc1201aeef96ca71b897dcdf3672f870c
Change-Id: I5c41decd4b89c5fee435cba6ae2d61f88b901900