[roll] Roll fuchsia [sherlock][i2c] Publish LCD bias module for testing.
The LCD bias module is not configured by the operating system so it was
never published by the i2c driver. We need to probe this module for
testing so we expose it on the bus now.

Test: device-enumeration-test
i2cutil r /dev/class/i2c/007 0x03

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584156
Original-Revision: 668ba9c05ae45e2e4778df68418737ded8360236
GitOrigin-RevId: 7dfd911e65b42013bcb9fa173707e1026a5cee54
Change-Id: Ifaf7b655bc2c70fc8915349be52d6b4dbb2a4862