[roll] Roll fuchsia [audio_core] temporary 4->2 channel-mapper
As a short-term device-specific workaround, the SourceReader class
that provides channel mapping from 4-channel to 2-channel has been
changed to simply copy channels 0 and 1 as-is, ignoring channels
2 and 3. Previous behavior was to emit the average of input channels
0 and 2 as output channel 0, and the average of input channels 1 and
3 as output channel 1.

We expect this workaround to be removed once the driver is updated for
the device in question. This is tracked with fxb/85201.

fx test audio_mixer_unittests

Original-Bug: fxb/85201

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584501
Original-Revision: 4ef3f365dbb3c2614afed9fb45362b24ad76fd6a
GitOrigin-RevId: 63a5bc1339ce241078159f076140ca728a78c440
Change-Id: I92786e149670ebd91f84a741d536679e01d0c021