[roll] Roll fuchsia [wlan][policy] addition of roaming framework
This adds structure for roaming to happen. When there is an ongoing
connection, we will get periodic connection data and then use that data
to decide whether to scan for a network to switch to. Only one scan
should be requested at a time. After this CL, a scan should never
happen; a future CL will add the logic to decide whether to do a scan.

Original-Bug: 83744
Test: Added a unit test:
	fx test wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579055
Original-Revision: e390741c649748f422b28d920386b276af794daf
GitOrigin-RevId: 003b71185e7af2ccb14861422f9f8d0a8a7601ad
Change-Id: I4c70812f289e78494782119b1be3ff5438e62457