[roll] Roll fuchsia [fvdl] Enable FVD manifest in ffx emu
As described in go/emulator-in-ffx-design, we aim to improve the
developer experience around the Fuchsia Emulator by defining devices for
testing using a Virtual Device specification similar to Android's AVD.

This CL adds the plumbing around a virtual device specification
manifest: a way to define a virtual device in a JSON specification file,
input that file to `ffx emu`, and propagate the values in that
specification to the emulator. The parameters in the specification were
all previously hard-coded values or command-line arguments with default
values, and this CL does not change the default behavior if no
specification file is provided.

Currently, the ffx code reuses similar functionality in the
device_launcher executable, with an eye toward removing ffx's dependency
on device_launcher in future changes. Similarly, the code for parsing
the manifest is currently built directly into the ffx code, but will
eventually live with the rest of the generic manifest parsing code in
//build/sdk/meta/src.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581361
Original-Revision: 39aafc786f71b95f91ebc0e41ced4700bfaedbf5
GitOrigin-RevId: b0bba26e06a5f75495146dc2f21e970af6b4c65d
Change-Id: Ie1ce9cf9c177afc8f45b532563585f8d1c4a10ae