[roll] Roll fuchsia [ffx] Implement `--recreate` for `ffx component run`
If a component created using `ffx component run` still exists in
the topology, the `--recreate` flag will destroy it and recreate
it. This is an alternative to the `--name` flag, allowing devs
to reuse the same component instance name. It also removes the
need for developers to call `ffx component destroy` manually.

Test: fx test ffx_component_run_lib_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584901
Original-Revision: 5f6986dc9679afa27e47901167051550683c2861
GitOrigin-RevId: 23e7e8f7efcb0136e1d6867ed082b58854c7e8a6
Change-Id: I21748d3dbdc09c03ceac23a90bd221186bb5ae63