[roll] Roll fuchsia [netcfg] Remove devfs_path arg
The path to devfs no longer needs to be configurable because it is the
same in netemul test realms as it is in production.

Also remove the netcfg-netemul component which was the sole remaining
usage of the argument, which is now unused.

Original-Bug: 78610
Test: fx test netcfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584906
Original-Revision: b413a62d52e030df16ed84c2ff56bc2b4e6d54de
GitOrigin-RevId: 9a8a2b7852499da0a8a18e8702d7c0330bff9d8d
Change-Id: I5ac40b5c997349890b24f3afe56d52b77a41a8a7