[roll] Roll fuchsia [paver] Additional logging for failed paves.
Currently the paver appears to be giving up due to premature file
ending. This information will confirm that by showing that the metadata
expects more left in the extent and yet the end of file is seen. Also
adds to the PayloadStreamer to show if it exits with leftover data.

Original-Bug: 83638

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585161
Original-Revision: af53905c6262918e8eddd05c66b0f89d6810d038
GitOrigin-RevId: f6951a3b68443d4617459b4c5c07db17d51417e2
Change-Id: I585b9331b84ab60eec5200c469873c2d46d323b6