[roll] Roll fuchsia [fshost] Inspect data for minfs rewrite progress
Adds progress indicator inspect nodes (which are exported at
bootstrap/fshost:root/minfs_upgrade:*) to let us monitor the progress of
the in-place minfs rewrite.

This is wired up to the fxfs rewrite for now to have it exercised
somewhere.  We should change it to trigger on the minfs rewrite once
that is ready.

Original-Bug: 84973
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584921
Original-Revision: f7818e686694bda7777133d76ce472a44861c75a
GitOrigin-RevId: 66ac6e44bd9999400058469a6157ff999935bf03
Change-Id: I37ee86c72a6f582b5efba74015031e14d0f08c93