[roll] Roll fuchsia [stress_tests] Stress Tests as TRF runner
This CL proposes a new test framework runner for
stress tests.

This runner abstracts away the test loop and allows
the actors to be written as components in multiple
languages.

In this model, a stress test is a component realm
with the following topology:

                 Test Root
		 /        \
       Test Subject       Actors

Actors are components that are created dynamically
by the runner in a collection. The test writer will
route capabilities from the test subject to the actor
collection. The actors will use these capabilities to
stress the test subject.

The actors are controlled by the runner using a
FIDL protocol: fuchsia.stresstest.Actor.

Actors offer a set of actions to the runner. The
runner coordinates between the different actors,
making each one run different actions in parallel.

The Test Runner Framework does not see this
internal implementation. Stress tests have a single
test case exposed to TRF that will error out when any
of the actors fail.

Multiply: stress-runner-integration-test
Test: fx test stress-runner-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567861
Original-Revision: 4f8b96e94898d4ab16c784a0cfe6a915e389570f
GitOrigin-RevId: d58ca16d9c8a863a49a9ae2f8530ce5e591545f5
Change-Id: I53d4423f7d7f6755ce7600af536138ec6444f0f4