[roll] Roll fuchsia [usb-cdc-ecm] Reduce "ignored interrupt" log message severity.
This can be triggered under some circumstances for some adapters. In
particular, some rt8153 adapters seem to start sending these spurious
interrupts when the device is unexpectedly rebooted but the adapter
remains powered (as is the case when using an externally-powered hub).
These interrupts don't seem to impact the functionality of the device
but the error message makes logs and the serial console virtually useless.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585903
Original-Revision: 20cb3dfd073bb02c510dc83083488640bd7d4587
GitOrigin-RevId: 1f47e4d8a16e91e446f6ca2aa69bff3448a7703a
Change-Id: I10328cb86ae1f101e6f6b5876309f44bee92ec93