[roll] Roll fuchsia [hypervisor] Rename zx_vcpu_resume to zx_vcpu_enter
This change does the following:
1. Renames zx_vcpu_resume to zx_vcpu_enter
2. Renames the kernel internals to match the change in name
3. Updates the associated tests to match the new naming
4. Updates documentation and comments to match the new naming

For the kernel internals, the choice was made to remain consistent all
the way through, as we have in the past for the implementation of the
hypervisor.

This is the first part of the implementation of RFC-0128.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585904
Original-Revision: 3f3495c5fd32963a59fa6887e78461170b29041a
GitOrigin-RevId: 30e8cf029a2418258d0e2d366dfddf7963141e3d
Change-Id: Iab108554a9f676c57b23a55818a2ab0bde897366