[roll] Roll experiences [ermine] Removing loading indicator upon view close.
If a view is taking a while to load, we show the loading indicator. If
the user close the view using keyboard shortcut before it finishes
loading, the shell gets stuck in an invalid state, thinking the loading
is not over.

This change resets the flag if the view is closed while it is still
loading.

Original-Fixed: 85143
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/585538
Original-Revision: c9bb15dd80648be21c17e17bf43ee4b6341c8532
GitOrigin-RevId: d1e5ff0d7524dcc4ca9611224536fdb4c35a034e
Change-Id: I34260b2ede9b394b50e7954809214d1e562ebdaf