[roll] Roll fuchsia [fidl] Fix overview comments
To have the description shown, it must be triple slash. There can be
only a single description on a `library` stanza, and per the rubric this
should be placed in `overview.fidl` if there are multiple files in a
library.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583983
Original-Revision: 3be4c79214cd3f677b8f16257cd32152187d12f7
GitOrigin-RevId: c263001e41de92e2152b73046d7ac586564768fb
Change-Id: I423792776a996c7e3fa10563dc6b58fe2ca5cb11