[roll] Roll fuchsia [fidl_c] Make fidl_validate an internal function
fidl_validate is currently part of the public FIDL API, however it is
intended to be used only by HLCPP internally to validate messages. This
CL renames the function to make it clear that it shouldn't be depended
on.

Unfortunately, there are two usages of fidl_validate that were left in
because there is not currently any alternative public HLCPP API.
- One validates HLCPP types for a test and could be transitioned to
encoding with Encoder::NO_HEADER with wire format specified, but that
also uses an internal API.
- The other validates event bodies and doesn't have a mechanism to
access the type

Original-Bug: 63427
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585443
Original-Revision: 520476a64af7668839725d0938a88112263a5ac9
GitOrigin-RevId: 6a680f5bd10887a769250b5e04d67fa274c9c0f7
Change-Id: Ia1f1b9953015f0386cae769278fef751298653a0