[roll] Roll fuchsia [ffx] Heuristic to find SDK root.
If we're working out-of-tree and the sdk.root config setting is unset,
we can try to find an SDK manifest as a parent of the `ffx` tool's
directory and use that directory as the SDK root. This will obviate
the need for most users to set this config setting.

Original-Fixed: 84342
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586489
Original-Revision: dc6cb3ef6136c72df54bd32b9442d7954bdb9c70
GitOrigin-RevId: 8f77b79e17e878a4c4b08aa82e3f76487bb77384
Change-Id: Idace320fb1ab0e829c54a885d18de0ba6a28dc28