[roll] Roll fuchsia [a11y] Only reconnect a11y view if it was already initialized.
It's possible for the a11y view to get a ViewHolderDisconnected event
when the call to CreateAccessibilityViewHolder() fails. In this case, we
should not try to re-initialize the a11y view, since it can result in an
infinite loop.

TEST: manually verified that session_control restart still works.
Original-Bug: 85216
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584877
Original-Revision: e1b727a44365fd4b8d2e09a4fff291b45d77bf11
GitOrigin-RevId: 745fdbcb4176693ab6e5391b7a9bee7abc62043a
Change-Id: I66d887c99d4b852f0170fb6c902fc81ced384bdd