[roll] Roll fuchsia [bt][hfp] Peer Connected events are sent after connection completes
Before this, spurious connection results would be sent if a Peer
connects but does not make an L2CAP/RFCOMM connection for the
HFP Service Level Connection.

Also sets a timeout on an RFCOMM connection from the peer.
If the timeout is reached, the peer task is stopped.

Test: Added unit tests
Original-Bug: 82575

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585623
Original-Revision: a1474bad5f2079efcb1da1a056c332b1dd0251dc
GitOrigin-RevId: 04fc02f356e6c1ae015c32bb9674ae4145b91cc5
Change-Id: I7618ba42f55c535d1655abde196027f870e193dc