[roll] Roll fuchsia [FMS] Use a mapping file for product bundle paths
The user would have to pass in a list of product bundle paths
via a GN arg which was painful. This makes it so the user
passes a mapping file location that contains the name of the product
and the path to the product bundle for that product.

Original-Bug: b/198166472
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585490
Original-Revision: 379f7948439b60975dc4da5d1adf1f86152f5493
GitOrigin-RevId: d494e33d73de1f62792f24f5dfe449eea6cec5a7
Change-Id: I19e1f177e57b90105f55087ba24f11a269067c22