[roll] Roll fuchsia [fserve] Add opt-in support for ffx repository
This starts the transition of fserve to using the new package repository
server that is built in ffx. In order to ease the transition, this is
made opt-in so that it can be tested out in field to shake out any
issues before being switched to the default.

Original-Fixed: 84290

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583629
Original-Revision: ea018cd23300fc0077dcc52bb3f068b11ddae021
GitOrigin-RevId: 4cb283b182be37f12ffb2debdc12fdbb1d7a0aa3
Change-Id: I9e503d384812c50262cd440be83eed9dff04db32