[roll] Roll fuchsia [ffx][assembly] Don't hardcode number of minfs partitions in fvm creation.
This is part of a change to support a minfs "account" partition in FVM
that will be encrypted with a key derived from the user's
authentication factor. This is in addition to the existing minfs
"data" partition that is encrpyted with a per-device key.

Prior to this CL the fvm logic in ffx was discarding all but the last
minfs partition in the board_config.

Original-Bug: 582846
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586504
Original-Revision: c8c50ce2e80249fdab69e07aa4f44d1e5cd515d2
GitOrigin-RevId: f7ef14e3aaf1c6801d8ebaf6a8304b41799d7eba
Change-Id: I63fac167e7be28f69283e7b07209162b8d62e14a