[roll] Roll fuchsia [run-test-suite] Make name optional when verifying artifacts
The names of files are not guaranteed to be stable, although the
current implementation of the directory output uses static names for
each file. With custom artifacts, multiple components in the same test
may produce a custom artifact, so using a static name won't be
sufficient. To allow testing in this case, this change updates the
testing module to key artifacts by the artifact metadata instead of
filename. This allows verifying the artifacts while allowing
implementation flexibility in choosing filenames.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583624
Original-Revision: 6ff2d1b51981837e341e23aa65351cf7ba23037e
GitOrigin-RevId: 783c3ff3d6aaf5f91283f5aa069aabd0fde48109
Change-Id: I1aed800caf4e21c13a7e46ce19696492fe96e198