[roll] Roll fuchsia [fidl][cpp] Refactor front-ends of back-ends
This is a major rework of the way that the various C++ generating
fidlgens are invoked. The goals are simplicity and consistency which are
useful as we move towards more binding unification and extend our
generators to support the new driver transport.

The command-line interface for these fidlgens have been updated to be
simpler and more consistent. The flags are now:
  --json path to the JSON IR file
  --root directory to generate bindings under
  --clang-format-path path to the clang-format binary
  --experiment turn on an experiment
Because the HLCPP generator is in the SDK its command-line interface is
a de-facto API. There's support for the old command-line interface but
the arguments that are passed are transformed into the new forms.

The command-line interface utility type takes care of loading and
parsing the JSON IR, largely because the library name is needed to
validate that the legacy arguments can be safely transformed into
updated arguments. If & when we drop the legacy this could be separated.

Previously, the filename of each output file was passed on the
command-line. This not only resulted in unwieldy command-lines and
complexity in the build system, but it also seemed to offer more
flexibility than the tools actually offer. The generated bindings have
to make assumptions about the include paths of bindings generated for
dependencies but passing the wrong arguments could make this silently
fail.

Now output file paths are determined by the generation root (passed with
--root) and templates. Each file generated by a fidlgen has a symbolic
name like "Header", "Source" or "TestBase". The fidlgen main calls
Generator.GenerateFiles with the loaded IR and a list of these names to
generate. A template named "Filename:NAME" is called to determine the
filename to use and then a template called "File:NAME" is called to
generated the contents. This decoupling provides both more simplicity
and more flexibility. There's a new template utility function "Filename"
that given a symbolic name and a library will return the appropriate
name. This means that dependency includes look like:
  {{ range .Dependencies -}}
  #include <{{ . | Filename "Header" }}>
  {{ end -}}

Experiment flags are exposed to templates through the new
"ExperimentEnabled" function that given a name returns a boolean
indicating if that experiment was enabled for this invocation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585041
Original-Revision: add6a7990199b875e32a2c529bc71893186257ff
GitOrigin-RevId: f954fa42d62a76100c7ba4f40aab7e1eb245c016
Change-Id: Ib7a924811af6cb1070aacab2a4ab22b92dfe140b