[roll] Roll fuchsia [fidl][llcpp] Event sending error trigger binding teardown
Similar to the client side, sending errors in events should trigger
teardown since the server and client would proceed with mismatched
state otherwise.

Also similar to the client side, teardown is momentarily delayed for
peer closed sending errors. This is because it's a common pattern for a
client to queue some operations and then close their endpoint, and in
case of an error from the server sending an event back, the server
bindings should still drain the queued messages from the client before
tearing down.

To centralize the logic for teardown and this peer closed exception
rule, we introduce a |fidl::ErrorOrigin| type which describes whether
an error happened when receiving or sending.

Original-Fixed: 78990
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585289
Original-Revision: 78c16e1dcac37266858bc3ee89ffb7e73422d77d
GitOrigin-RevId: 5a07d34de95101f453914dd8c1a481d30805b2e5
Change-Id: Ib2208c8b16ecf39836598971012083d7276d8d68