[roll] Roll fuchsia [component][fuzzing] Add Runner class.
This CL introduces the runner class, which is responsible for performing
the details of various fuzzing workflows. It uses a dedicated thread for
these workflows to avoid blocking the FIDL dispatcher thread. It owns
and manages the corpora and the mutagen to generate inputs, and the
module pool to collect and analyze feedback.

Original-Bug: 24740
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567108
Original-Revision: 106c2bf8064d35b0244483293eaef11b99afaa95
GitOrigin-RevId: a7bcfe7e121cb439848ffa0a68c214d3fb1b6082
Change-Id: Ib353baab129d8bae212c491e4d627ae65ae7e366