[roll] Roll fuchsia [cmc][cm_fidl_validator] Storage capabilities aren't nodes in cycle detection
Currently, a storage capability declaration, like any other capability
declaration, counts as a `Self` node in the dependency graph.

This means that a CML like this results in a compiler error:

```
capabilities: [
    storage: {
        name: "data",
        from: "#backend",
        ...
    },
],
use: [
    {
        protocol: "fuchsia.foo.Example",
        from: "#child",
    },
],
offer: [
    {
        storage: "data",
        from: "self",
        to: "#child",
    },
],
children: [
    { name: "child", url: ... },
    { name: "backend", url: ... },
],
```

due to an apparent cycle between `self` and `#child`. However, this is
an illusion: the storage capability is only an intermediary and is not
hosted by "self", but is backed by a directory exposed by `#backend`. In
other words, it is safe to stop `self` before `#child` because `self`'s
runtime isn't actually supporting any capabilities used by `#child`.

Instead, draw an edge directly between the target and backing dir in the
storage decl.

The end result of this change supports a nice property. For any
three-component topology of the following form:

```
    non_executable_component
           /     \
          A       B
```

There is an isomorphic (in terms of dependency order) topology with just
two components, that transforms offers from A to B to `use from A at B`:

```
    B
     \
      A
```

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584370
Original-Revision: c22cd5550f04948d5466c66d5ec1384fce1a9bc2
GitOrigin-RevId: 77a6bde485cded71374e350157ca29ffef794468
Change-Id: Iad33f0e2f7287113d42335e7e23f6924a63ced5e