[roll] Roll fuchsia [scrutiny][core] Tag components that come from static pkgs
This change makes the Scrutiny core plugin, which collects packages and
components, depend on the collector that to retrieves the static
packages list. This allows the core plugin to, when the static packages
list is available, tag components that come from a static package, allowing
controllers to verify that certain components come from trusted sources such as
ZBI and the static packages list.

Original-Bug: 81537
Test: fx test scrutiny

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582541
Original-Revision: 9c85343f10b268933089601f66933de882f4a7a6
GitOrigin-RevId: 3409ef27e5df6bae2f81d5b67e03757e640deca2
Change-Id: Ie2fb4c8a88ee05f6076eb080197ada5c69133d48