[roll] Roll fuchsia [component][fuzzing] Add Transceiver.
This CL adds the Transceiver class, which is used to convert Inputs to
and from Zircon sockets that can be exchanged with off-device code, e.g.
an ffx plugin via overnet. This class is responsible for transferring
file-like objects to and from the host. It uses a dedicated thread so as
not block either the FIDL dispatcher or runner when transferring
arbitrary amounts of data.

Original-Bug: 24740
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567109
Original-Revision: 18a6b061ca93c967d008a71492661c1aac850d5b
GitOrigin-RevId: dba76ca1aea45c183fecc8e636421cd92092b2ac
Change-Id: I404172e87af95d9d56f989e9a6519aed8c7bc8c6