[roll] Roll fuchsia [fidlc] Don't overwrite unchanged files
Before overwriting files fidlc checks to see if the contents match. This
significantly reduces the amount of work done by the build system when
there are changes made to FIDL source files that don't affect the
output.

The check first checks for existence, then compares the sizes of the
files and finally reads the existing file and compares the contents.

Original-Bug: 66684
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586601
Original-Revision: 12d9e55fae34406baf75179819cc7432cd649368
GitOrigin-RevId: 4da15ce82502858d1d2ce391dcb0316fcdc199f0
Change-Id: If82fd7528c8ada6ab7af610f8da9551c09d32578