[roll] Roll fuchsia [fidl][llcpp] Fail TryDispatch early for invalid message
WireTryDispatch tries to dispatch the message against a server
implementation, and doesn't consume the message if a handler is not
found. This is used to support ad-hoc composing several protocols by
trying to dispatch one after the other in a chain of TryDispatch.

This CL is a collection of some improvements:

 - If the message has an error, exit the chain of TryDispatch early
 since it is meaningless to keep trying otherwise.
 - Mark the return value [[nodiscard]] which also revealed one usage
 that forgot ignored the return value.

It would be ideal if we could remove TryDispatch since it subverts the
FIDL protocol semantics. However, there are still a few drivers that are
load-bearing on it. See fxbug.dev/85473.

Original-Bug: 85473
Test: fx test -v //src/lib/fidl/llcpp

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586486
Original-Revision: 1d4d842088d4de974ed2876c202d9c673539e247
GitOrigin-RevId: 737fab553d1438db30ffd72cd2a09ebc3656aa1d
Change-Id: I289197e362a2ec9d868fb65db67228bdcaaded38