[roll] Roll fuchsia [virtualization][vmm] Decode OR instructions with opcode 80 and 81
Clang generates OR instructions with opcode 80 for virtio PCI backend
code. This type of instruction reads data from memory, performs an
inclusive or operation between the data and an immediate value, than
writes the result back to memory. This CL adds support for decoding
instructions with opcode 80 and 81, so that we can cover all cases where
an OR operations is performed between an immediate and a block of memory
with the same size.

Original-Bug: 12611
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576731
Original-Revision: ca98a4bf345d003974b3afe1c227c4825e1ecec1
GitOrigin-RevId: b283901cc351f590a3d7ac3e28e4cce29e3cb314
Change-Id: I7f750a3f0beefa0cb12cf8aca057e11968871c71