[roll] Roll fuchsia [cm_fidl_analyzer] Build ComponentTree within ComponentModelForAnalyzer
This is a transitional change that prepares for Scrutiny
to use the new version of the capability route static analyzer.
That new analyzer is based on the ComponentModelForAnalyzer
data structure, which wraps the ComponentTree structure used
by the old version. This CL:
- provides the data needed to construct a ComponentTree to
  ComponentModelForAnalyzer, rather than providing the tree
  itself,
- constructs the root environment based on the RuntimeConfig
  passed to ComponentModelForAnalyzer (currently just a
  placeholder function) and propagates it to environments that
  inherit from root.

Original-Bug: 61861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584845
Original-Revision: f4ae30d7b8a177489bb12789b62b9a34b4118bb9
GitOrigin-RevId: 3e8a41becd2c39c447602296b33f0f2a7b4263a3
Change-Id: Ic8edade14e10f9455d73b9d0eae2cb34c6588f3e