[roll] Roll fuchsia [component][fuzzing] Add Controller class.
This CL adds an implementation of fuchsia.fuzzer.Controller for the
fuzzing engine. As the management interface of the fuzzer, it interacts
with the Runner to perform fuzzing workflows, and with the
Transceiver to transmit and receive file-like objects to and from users.

Original-Bug: 24740
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/558390
Original-Revision: 7dcc415103325530d7d4dd60bd00eae03b66fc8c
GitOrigin-RevId: 5b785a854d8ed81856e85b4dce28c7b744530a31
Change-Id: Id0a5265532967c0ea98682c90e8d3dea849bb24a