[roll] Roll fuchsia [fidlgen][llcpp] Simplify decl iteration
Our fidlgen templates have a lot of passages like:
{{- range .Decls }}
{{- if Eq .Kind Kinds.Const }}{{ template "ConstDeclaration" . }}{{- end }}
{{- if Eq .Kind Kinds.Protocol }}{{ template "ProtocolDeclaration" . }}{{- end }}
{{- if Eq .Kind Kinds.Struct }}{{ template "StructDeclaration" . }}{{- end }}
{{- if Eq .Kind Kinds.Table }}{{ template "TableDeclaration" . }}{{- end }}
{{- if Eq .Kind Kinds.Union }}{{ template "UnionDeclaration" . }}{{- end }}
{{- end }}
where we go through all of the declarations in the library and then for
some subset of declaration kinds we dispatch each to a different
template.

This CL introduces two new kinds of method on the root C++ IR object:
 - methods for each kind of declaration that return all of of that kind:
   .Consts, .Bits, .Enums, etc
 - .ProtocolForTransport that returns all protocols supporting a given
    transport.

All of the templates for llcpp have been rewritten to use these methods
instead of .Decls. The result is that the order of types generated has
changed but there's no semantic difference and I think it's more
readable.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587249
Original-Revision: 3f785af91832a545eaf4b79b3c1c37364a29e0cf
GitOrigin-RevId: bce85ac37035aaaaed66699c5bd9a1b5ffb658a9
Change-Id: I543ea86999dcc1efbb61f7e96456952d1cf299db