[roll] Roll fuchsia [rfc] Fixing RFC-0082 area
The `PosixCompat` was not listed in _areas.yaml, adding it there for
now. As noted in the TODO, we plan to align this list with overall
project areas, and can revisit this particular area choice at that point
in time. In the API world, we have the `fuchsia.posix` library prefix so
it seems sensible to have an API area dedicated to POSIX compliance.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587403
Original-Revision: 42cbe94efb88958276702304728c0bbf909b0954
GitOrigin-RevId: 410bc0bdf4ae0d7a2f926a36e173137a3881b1dc
Change-Id: Ie700b093c7800ca2e1b3905aa2bf3f3aaaea521c