[roll] Roll fuchsia [ffx] Rearrange self-test code
There was a question about adding tests to the self-test plugin. It
seemed like a bad idea to have a giant file that accumulates tests if
different plugins want to also put test inside there. I suggested
creating a module for each plugin to contain their own tests. I figured
it would be good to clean up the current tests to follow this pattern.
This should make it easier for people to add tests without merge
conflicts.

The only visible change is that the test macro stringifies the name to
the function to display so "test_daemon_echo" becomes
"daemon::test_echo" in the output. This shouldn't be too bad and it
actually helps organize the output a bit as you can see what the test is
roughly about first by the module name rather than having to read past
"test_" to get to it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584863
Original-Revision: 468f59a574e65460832867d460ea9ec503b18760
GitOrigin-RevId: 299e66de8c0a2a5b848727fa89a1cb5be7f515b3
Change-Id: I7a91d8f00de7ce39ee3b229d2ebe0a97d17a507e