[roll] Roll fuchsia [ffx] Add self-tests for config
This adds some simple env, get, and set tests for ffx config. As ffx config is
one of the top user flows it is desirable to have tests for it. This can
be fleshed out more in the future, but that will be easier with
something in place.

Original-Bug: 82378

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584864
Original-Revision: 77d228fe5605ea445604478fde00416ba3cdf588
GitOrigin-RevId: a5d133971f096866dbf02e8460fdf2781eec7d93
Change-Id: I43d9150278b631d34a218bd8dea4063aa2398e9d