[roll] Roll fuchsia [codesize] mark library as not null safe
In order to allow our build rules to specify that a program
should compile with sound-null-safety we need to check if all the
libraries are marked as null-safe. This library is saying it is
null safe but opts out of all but 1 file which breaks the build
rule check. This CL will opt out of null safety but then opt in
for the one file that is using it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587792
Original-Revision: 3e9a12cb1171d807c8e2099b5d7adf840a18fc29
GitOrigin-RevId: f5fc99fa11f262430dd29f5b1d2532f0ca38cd86
Change-Id: I2a3f2ccf621685819e6147bcb8ecbc43c2d38d72