[roll] Roll fuchsia [fshost] Format minfs when writing fails during a resize
Write a file to minfs to indicate when writing the contents of the old
instance to the new instance has started. This file is used to detect
if the device was restarted during the copy operation which could
potentially leave the device in a unknown state. If the file is detected
after a restart then minfs is reformatted.

Overwrite the minfs magic before removing the partition to prevent
mounting minfs on the next boot if deleting the partition partially
fails.

Original-Bug: 84885
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586862
Original-Revision: d2e282101de919142a4a1c751dc198fc0766d7e6
GitOrigin-RevId: a0445207b2927cd17abb03ca63dd957216f4abf0
Change-Id: I902e517545a691526a8f856a946e6d1df92e0ae6