[roll] Roll fuchsia [fx] Make new-sync-to support download snapshots from GCS
In replacing the legacy internal `fx sync-to` command with a new
public version, we must continue to support syncing to every release
branch that's still under active development. That includes branches
from before recipes started emitting the `checkout_info` output
property that `fx new-sync-to` currently uses for determining how to
reproduce a checkout.

Therefore, if a build doesn't have the `checkout_info` output property
set, `fx new-sync-to` needs to replicate the legacy sync-to command's
behavior of downloading a Jiri snapshot that the build uploaded to GCS,
and syncing each repo to the revision pinned in that snapshot.

I tested this on some old release branches to make sure the
functionality works and is consistent with the legacy `fx sync-to`
command.

I also updated `fx repro` to suggest using `new-sync-to` for builds that
the legacy `sync-to` doesn't support.

Original-Bug: 68492
Original-Bug: 72533
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587621
Original-Revision: a7a94a4808d77aaa558c6e77c8aa778cb236e4be
GitOrigin-RevId: 1bb80bed2faf5906fe09cafb40a7723a93c67eb3
Change-Id: I99272cff2b49e44cb432f93ee31faacd5ba17ca3