[roll] Roll fuchsia [mediaplayer] Fix audio renderer issues when used with AudioConsumer
AudioConsumer clients are allowed to start the timeline with SetRate
without specifying a subject time. This can lead to the AudioRenderer
being started on the incorrect media time, which causes
silence/stalls/glitches.

Add another state to FidlAudioRenderer to allow AudioConsumer to
defer calling Play till it's user has called Start with the real
subject time. Allow packets to be sent to the AudioRenderer in this
state to 'prime' it, and prevent stalling the upstream StreamSink.

Test: fx test mediaplayer_tests
Original-Bug: 83451

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587646
Original-Revision: 4b1cd94821d4cc6b8c4289ed21de249533b55667
GitOrigin-RevId: 476f4424ff45cac7cb6b4f6db380d46a336bf037
Change-Id: Ic3e89235193510fe09ad931e6d9f0e363c5311fb