[roll] Roll fuchsia [bt][emulator] Manipulate emulator from DirectoryProxy
The DeviceWatcher, Emulator, and TestDevice utilities all operate under
the assumption that the Fuchsia Device Manager's top-level `/dev` dir
is accessible from the namespace of the component in which they run.
However, when we start an Isolated Device Manager within a RealmBuilder
in the current component, the isolated `/dev` directory is not in the
namespace. It is explicitly routed out of the Realm. As such, we change
these utilities to allow for passing in an already-open `/dev` dir in
which they can manipulate the emulator.

Multiply: hci-emulator-client-tests (fuchsia): 100
Multiply: bt-device-watcher-tests (fuchsia): 100
Original-Bug: 71895
Test: Added `tests::publish_lifecycle_with_realm` to hci-emulator-
      client-tests, which uses the same expectations as the (currently
      flaky) `tests::test_publish_lifecycle`, but with an IsolatedDev
      Mgr in a RealmInstance instead of using the system DevFs.

      Rewrite currently-ignored bt-device-watcher tests to use the new
      directory-relative utilities and an IsolatedDevMgr inside a
      RealmInstance, and add one to verify the new relative-path
      handling WAI.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552767
Original-Revision: c7f5881507a880145fcc1d642bf0b33089c7fe4a
GitOrigin-RevId: e384a2db4ef28876c1a7ac190f7ea098fc956a7d
Change-Id: Icf7e9cb4d12afb46f4a70e1892bda3f4689e5824