[roll] Roll fuchsia [bt][hci] StopAdvertising() called immediately after StartAdvertising() should result in advertising disabled
If StopAdvertising() is called immediately after StartAdvertising(),
nothing happens, and the StartAdvertising() call successfully finishes
and the device will be in an advertising state.

This change updates StopAdvertising() to result in the device being in a
non-advertising state, regardless of whether a previous
StartAdvertising() call completed yet or not.

Original-Fixed: 50542
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575790
Original-Revision: f2a52c34e4317661be10f7dc13d93898cd3e8934
GitOrigin-RevId: fa94533882751e634e3f5291af14d306315fc9b5
Change-Id: I6413bcbd3bcf31687e0e21a411c7f3c147b0a929