[roll] Roll fuchsia [virtualization][vsock] Simplify Mux/Demux control flow
The VirtioVsock::Mux and VirtioVsock::Demux methods are called by
a VirtioQueueWater object when descriptors become available on the RX/TX
virtio queues. While in general, the VirtioQueueWaiter is convenient for
virtio code (because it passes the index of the next free descriptor to
the callback), in the case of VirtioVsock it complicates the code
a little: for example, the Mux/Demux functions don't always want to use
a descriptor (and hence need to return it unused).

Replace VirtioQueueWaiter (and the associated wrapper class "Stream")
with just a plain async::WaitMethod object.

Multiply: virtualization-core-tests, vmm_unittests
Original-Bug: fxbug.dev/85702

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584721
Original-Revision: 7b24b46e3ade70052e9073c9e723df5517e8c5a7
GitOrigin-RevId: 523ca62a4ceefae487c5e16bfe6e0c6580eb1b21
Change-Id: Ic1500de42daf69e5620ec8945ca002dfff1a1966