[roll] Roll fuchsia [events] Add EventSequence.has_subset
The has_subset sequence matcher allows matching against a subset
of events in the stream. Unlike all_of, the EventSequence will
not return an error if it sees an event that does not match
a matcher in the group.

This is useful if the event stream contains unrelated events, such
as those created by a different test within the same component.

Original-Fixed: 85547
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586603
Original-Revision: 436e47e789ea4831895178b981bfb67300f100b5
GitOrigin-RevId: 400a1b318aef47c5f8342d33d4a3e2dc8211f3e8
Change-Id: I700178b8ef88f99a454780af78a701359dfeb797