[roll] Roll fuchsia [ffx][repository] add-from-pm should default repo name to "devhost"
This incorporates feedback from UX that wants us to both standardize on our
syntax for specifying repository names, and to simplify the arguments in the
simple case where a user is adding a single repository. This switches
`ffx repository add-from-pm` from taking the repository name as a positional
argument, and over to having it be an argument `--repository $REPO_NAME`.
Furthermore, this changes the repository name to default to "default".

Original-Fixed: 85199
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588191
Original-Revision: 070d7ae2d393afdd4cf5dfb049e05417823873af
GitOrigin-RevId: 126ae5c2da10078d25a4e826dce46c87ce949b64
Change-Id: I277c529eebf79f08c4123116410898b1561d1df1