[roll] Roll fuchsia [bt][avrcp] Use PSM from search result for outgoing connections
It's possible for some devices to advertise a PSM for an AVRCP service
other than PSM_AVCTP. This is not likely, but the peer may expect us to
connect to it.
Update AVRCP to parse the PSM from the primary protocol of the BR/EDR
search result. Use this PSM to make outgoing connections.

Original-Fixed: 63715
Test: fx test bt-avrcp-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581384
Original-Revision: 7efe5e7bca8a304d4160c66133483b5805a3b182
GitOrigin-RevId: c0dfee729981a1012dd95ec7016ffef2ed0cb979
Change-Id: I69ef2fbb8441e8c8959cacf7b1970680388907b6