[roll] Roll fuchsia [fidlc] Restrict @transitional to methods, @unknown to enum members
This CL removes a lot of complicated attribute validation around
@transitional and @unknown for enums, bits, and unions that was never
used for anything in backends. Now @transitional is only allowed on
methods (not on types as a way to allow @unknown despite strictness),
and @unknown is only allowed on enum members (not bits or unions).

Test: fx test fidl-compiler
Original-Fixed: 84334

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587909
Original-Revision: 9720964822189d0f27763ea485ccca7e1179d2de
GitOrigin-RevId: f668c5c2ad23f745a3299cb61fbab726f3b1820a
Change-Id: I782f9aad5fcd65c4679cd3b49360157c1cd56b69