[roll] Roll fuchsia [test_manager] Stop running a suite when the controller handle is dropped
Fixes a regression where the suite is no longer killed when it's
controller channel is closed. This could cause test manager to crash if
it attempts to send a suite event when the future serving the controller
is already finished.
This works because unlike an AbortHandle, a RemoteHandle cancels the
future when the handle is dropped.

Testing: fx test test_manager_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586171
Original-Revision: 9833860789fe1c23435275e9519611c370c96a8c
GitOrigin-RevId: 62b4edfa3aa1bfc8c153c2ad553224f0123b58ba
Change-Id: I9ef5b0332b1ec9a52abb6f410ae57d01de4ed2fe