[roll] Roll fuchsia [kernel][vm] Consider aspaces used if modified
Any operation that modifies the mapping in an arch aspace can modify it
in a similar way to what is being detected by active_since_last_check_
flag. Not setting the flag when these modifications happens will cause
scanning to falsely think the aspace has nothing new to scan, when it
might. This CL causes any operation that changes the aspace mappings to
treat the aspace as if it has been active, and so might have new
information in it.

Original-Bug: 85512
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587302
Original-Revision: 1264aabd206de4b0c83d8b4c5f75eda28a7e3216
GitOrigin-RevId: 336b548e251eb6dde854750608f7bf23d696b09d
Change-Id: If3aeb093308dd50f3f51f2030f3966aad30a4ad3