[roll] Roll fuchsia [llcpp] Add ContainsHandle trait for num handles calc
This CL adds a ContainsHandle trait which is true if the object contains
a handle. This is used for computing the number of handles when envelopes
are inlined. The previous approach used erroneously didn't count structs
with handles as containing handles.

Original-Bug: 79578
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588421
Original-Revision: a1cc0da3feea2e63bbe5c0c1f23dbc2341a0e4b6
GitOrigin-RevId: fc5126fe6bfb11826c287218666d6def226ec142
Change-Id: Ia96a48954f97ca91f03504467dc9c27d09c69cfb