[roll] Roll fuchsia [wlan] Fix bitfield/element conversion warnings
A lot of the wlan codebase is built with conversion warnings disabled.
When creating a new build target with conversion warnings enabled some
of the common files and one of the unittests came up with warnings as
errors. Fix these by casting to the proper types.

Test: fx test wlan_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588227
Original-Revision: 236d2675449095d276de1718c21bc5a3920fa444
GitOrigin-RevId: 9000a06a8d0f4c0bcd5b8ff16af5825441eb72d0
Change-Id: I2e28228670402502562a690c6cbc71d61156fb31