[roll] Roll fuchsia [subprocess] Wait for subprocess to finish after sending SIGTERM.
The subprocess command like botanist or testrunner may have cleanup
steps that it should run even if it gets a SIGTERM, such as cleaning up
temp directories or killing an ffx daemon. This change sends a SIGTERM
(instead of SIGKILL) to the subprocess and waits for it to handle the
signal properly before exiting. This is needed for using ffx in infra to
ensure the ffx daemon gets stopped properly even if testrunner gets
terminated early.

The logic is copied from
https://fuchsia.googlesource.com/infra/infra/+/d3b5c7a7a4bb0d4182cd57be72acd31462668e0f/cmd/catalyst/main.go#31.

Original-Bug: 84869
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588281
Original-Revision: 114722ddc47bba23af8a1f09894cbf26f4bbde9c
GitOrigin-RevId: 59cd7351c33a6ede2c7e449f09f8c69d52fc2d88
Change-Id: I0791a76094f7ba6f6608ca93eff97936481ed800