[roll] Roll infra/recipes [checkout] Emit patch base revision for `fx sync-to`
Because we use `jiri patch -rebase-branch`, we always patch on top of
whatever commit is at the tip of the branch that the patched change is
targeting, which will vary depending on when we run `jiri patch` and
isn't currently recorded anywhere. So to ensure `fx sync-to` knows the
correct revision to patch onto when repro'ing a tryjob, we must record
and emit the resolved revision that we patch on top of.

Original-Bug: 68492
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/588782
Original-Revision: cbc435ba71d8c9331f64b98e0174dda38d2cb336
GitOrigin-RevId: 2825d5879d1ed9618426996f338e42d7bfff617b
Change-Id: I34e7fc81f1b4f125a8666673287e61b8553e303f