[roll] Roll fuchsia [dart] set HOME in update_3p_packages.py
A recent change to pub requires that the HOME variable be set
to cache some values. We do not want to rely on this variable
being set so we set it to a temporary directory for the duration
of our subprocess.

https: //github.com/dart-lang/pub/issues/3167
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589164
Original-Revision: 543b8faeed6a0602a0f93cf7e9ccb7dd8d0c04d2
GitOrigin-RevId: 7d746917a9317dacf8d570e499e288849fe0c0a2
Change-Id: I153a59f6251717ca91e39fd1386a8ab7b95b3177