[roll] Roll fuchsia [component][fuzzing] Add Dispatcher and improve Binding
This CL promotes fuzzing::FakeDispatcher to fuzzing::Dispatcher and
replaces usages of async_dispatcher_t pointers with shared pointers to
Dispatchers instead. This allows the async::Loop to be encapsulated, and
the thread it starts to be available to fuzzing::Binding. This allows
the latter to correctly determine when it is running on the dispatcher
thread and perform tasks directly, vs. when it should post tasks to the
dispatcher and wait.

Original-Bug: 24740
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585484
Original-Revision: 31891cd0a14b124be2ff1f9fc412b5b43b582dcc
GitOrigin-RevId: f60d8e958ff2114e65dc22d0e060a35b2ac1a17f
Change-Id: I1f120c9300bdde750e371690dc03c0e775e5f4e3