[roll] Roll infra/recipes Reland "[checkout] Emit patch base revision for `fx sync-to`"
This reverts commit 3b815a80ec930ba01646c2b46be91ec53d43c9bb.

Reason for revert: the original CL broke CQ because `jiri project`
returns '.' for the relative path of a project that's the root of the
workspace (as is the case for fuchsia.git), and that broke the recipe
engine when trying to do `api.context.cwd.join(path)`. This didn't show
up in recipes.git CQ because we don't patch fuchsia.git changes in
recipes CQ tryjobs.

The fix is to use the `path` value, which is absolute, instead of
`relativePath`, from the output of `jiri project`.

I also updated the default mock output of `jiri project` to indicate
that `path` is absolute.

Tested with led this time:
https://ci.chromium.org/swarming/task/5656eff4c662dc10

Original change's description:
> Revert "[checkout] Emit patch base revision for `fx sync-to`"
>
> This reverts commit cbc435ba71d8c9331f64b98e0174dda38d2cb336.
>
> Reason for revert: breaking CQ
>
> Original change's description:
> > [checkout] Emit patch base revision for `fx sync-to`
> >
> > Because we use `jiri patch -rebase-branch`, we always patch on top of
> > whatever commit is at the tip of the branch that the patched change is
> > targeting, which will vary depending on when we run `jiri patch` and
> > isn't currently recorded anywhere. So to ensure `fx sync-to` knows the
> > correct revision to patch onto when repro'ing a tryjob, we must record
> > and emit the resolved revision that we patch on top of.
> >
> > Original-Bug: 68492
> > Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/588782
>
> TBR=atyfto@google.com,olivernewman@google.com,auto-submit@fuchsia-infra.iam.gserviceaccount.com
>
> No-Presubmit: true
> No-Tree-Checks: true
> No-Try: true
> Original-Bug: 68492
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/589021

Original-Bug: 68492
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/589022
Original-Revision: 596fe0a9b636d5865ad03a4cc2c0f999d9f12d5c
GitOrigin-RevId: 61d814c84add6deebe96cd027b0683f20c4792bd
Change-Id: I698cdcc99b1244027149ed8888b7d4bc233b5811