[roll] Roll fuchsia [packet-formats] Better IPv6 protocol number logic
In `Ipv6PacketRaw`, store the body and protocol number together since,
if extension headers can be parsed, they are both present, and if not,
neither is present.

In `Ipv6PacketBuilder`, rename the `next_hdr` field to `proto` and
clarify that its value is the upper-layer protocol number, not the fixed
header's Next Header value (though they happen to be the same if no
extension headers are present).

Test: fx test packet-formats -- \
      test_packet_builder_proto_not_next_header

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585083
Original-Revision: 1211b661aa87995f4295136e83ed5a639804b567
GitOrigin-RevId: 6a289fcc27252258e47b7bc3bb8dcf765a04e335
Change-Id: I759eba4bda818ba468c4a7fe9d3026d4b492c2fa