[roll] Roll fuchsia [ffx] Add the ssh key to the self-test environment
I *think* there are test flakes due to ssh issues. I wrote a test for
component list which was always failing with a timeout trying to reach
the device. I spent quite a bit of time debugging and finally tracked it
down to an issue where the ssh.priv key in the config does not resolve
to a file so the daemon/overnet ssh code fails when it tries to run the
host pipe. What I don't understand fully is why these tests ever
succeed. It appears that the target tests do not actually use the RCS
and instead use a separate proxy and so the fact that RCS communication
is not working does not matter for those tests. But I don't really get
that story completely yet.

Anyway, this adds two mechanisms for setting the ssh key. One is used in
CQ which is an environment variable FUCHSIA_SSH_KEY which is also the
first thing that gets checked in ssh.priv. If that is present in the env
then we add that to the isolate subcommand. You can thus also set that
yourself locally if you want to use that route. I also added a command
line flag which can be used when running these tests manually rather
than in CQ. You can use that to pass the right file path.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585290
Original-Revision: b092f4bbda11299b0df0f000dd1bf7de8fb6c1b1
GitOrigin-RevId: ce13a6f3f73eff46fe1e2174c710e2ddb14abd7e
Change-Id: I0066af58f8a8c0de522187632735180acc18344f