[roll] Roll fuchsia [x86][acpi] Refactor ownership a bit.
Previously, ACPI was initialised as follows:

1. The X86 driver proper creates an acpi::Acpi instance, which it passes
   to publish_acpi_devices().
2. publish_acpi_devices() creates an acpi::Manager on the stack, which
   does device discovery and publishing.
3. Each ACPI device was given a pointer to the Acpi implementation, and
   created its own async loop (and thread) to handle FIDL requests.

In these circumstances, the Manager is destroyed as soon as all devices
are published.

In preparation for upcoming ACPI event work, this CL refactors the
ownership of various objects to be like this:

1. The X86 driver creates an acpi::Acpi and an acpi::Manager. It passes
   the acpi::Manager to publish_acpi_devices(). The acpi::Manager
   contains a single async::Loop for all ACPI devices.
2. publish_acpi_devices call the discovery/publish methods.
3. Each ACPI device is given a pointer to the acpi::Manager.

This means that we have less FIDL threads, and that future work that
involves sharing state between multiple devices can be done in the
acpi::Manager.

Original-Bug: 85785
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588601
Original-Revision: 1dc1b8290bceff636dcb90725e31f00b091afd3b
GitOrigin-RevId: 8c1d4c49f4d3f208afdc5085e1eacc644b5e8ef0
Change-Id: Ieb76083fe7549a9dd62a51121bbf6b9afe9453b2