[roll] Roll fuchsia [ffx] Ensure self-test cleans up daemons
This adds some code to the self-test runner to ensure that any daemons
are killed before finishing the tests. This is defensive and will cause
the test suite to fail if we did have to kill a daemon but should stop
the behavior in the attached bug.

Original-Bug: 76646
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589405
Original-Revision: 5c30e46904761d79c3560135e5bf3ada5e9388a4
GitOrigin-RevId: 7f27e90b16512c8d950155f493a5358ddeaa4ed1
Change-Id: I58c039e54953cd75e2f93dd050b5f6c393c037f7