[roll] Roll fuchsia [audio_core][test] Free memory sooner during test
When RendererShim objects are created, an additional reference
is held until the test case exits. HermeticFidelityTest did not
realize this and created a new renderer for each of the 42
frequencies that it checks -- expecting the renderer and its
members to be freed automatically when the renderer went out of
scope.

HermeticFidelityTest now expressly calls Unbind on each
renderer, freeing memory immediately. This saves approx. 50MB
of memory, over the course of a HermeticFidelityTest case.

The problem was first noted -- and verified as fixed -- with
the log messages emitted by audio_core as it pins critical-path
memory (including renderer VMOs) every five seconds.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588541
Original-Revision: 83a7f92c583234a20bb1ad382b058ec6c7fa1b95
GitOrigin-RevId: 3ab867b6feeaa7d6671a84821045668dab19298f
Change-Id: I117e8ec285eb852e7094511affd93957973e3fd0