[roll] Roll fuchsia [fx] Rename new-sync-to to sync-to
This new version of the command can be considered stable now that it's
passed all of my manual testing.

Internal checkouts will continue to use the internal version of the
command with the same name, but I'll soon deprecate it and rename it to
`fx old-sync-to` at which point all users will be switched to this new
version.

Original-Bug: 72533
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589946
Original-Revision: 1c4b4d5d5d89a5a54bd926334d99c5aed602f0d1
GitOrigin-RevId: 5b67f504b2978b19c503126de0ca05f7e9557441
Change-Id: I522092b768df802bca98478278b595d4d298b76c