[roll] Roll fuchsia [audio][lib] Default magnitude for non-float
The audio/lib/analysis method GenerateCosineAudio uses a default
magnitude of 1.0 if none is specified by the caller. This is
appropriate if generating audio in the float32 format but not if
the format is int16, uint8 or int24-in-32.

In those cases, the default magnitude is now that format's
maximum positive value. As an example, when generating int16
audio this method will now by default generate sinusoids in the
range [0x8001, 0x7FFF] or [-32767,32767]. As with any integer-
based format the most negative value 0x8000 cannot be used
without clipping since there is no matching positive value.

New unittests are dded to explicitly validate non-float defaults.

fx test -o audio-libanalysis-unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588561
Original-Revision: 7fa29ad9918385d390f321c002bcb73aea1f424b
GitOrigin-RevId: a65189543f816b2357c63eb08879f8d4cf27a497
Change-Id: Iefb00f958519fb1f4824aea182c7a5fe7b355edb