[roll] Roll fuchsia [fxfs] Use journal reset as a replay indicator
This CL addresses a bug during journal replay which results in valid
records being discarded.  During replay, the checksum validation logic
only verifies checksums since the last device flush (which is marked by
DidFlushDevice records), since (a) it's not necessary to validate
checksums which were flushed, and (b) after a DidFlushDevice, the
allocator is free to re-use a block for a new extent, meaning checksums
before that point are stale and not valid.

However, DidFlushDevice records are not guaranteed to be persisted each
time we flush the device.  We write them out with the first transaction
after a flush, but clearly that might never come before the system is
unmounted or has a power interruption.

Thus, during replay, we might not know about all of the points when the
device was flushed, and therefore might attempt to validate stale
checksums that preceded a flush, which can result in checksums seeming
to be invalid and records being discarded.

To address this bug, this CL makes three changes:

- Previously, the journal reset mechanism was used to indicate that the
  journal stream was partially written out (i.e. we didn't completely
  write the last block out).  This CL changes it to simply indicate that
  the journal was replayed, meaning the journal is reset each time we
  remount the filesystem.
- Journal resets now count as a device flush when deciding where to
  verify checksums from.
- When the filesystem is mounted in read-write mode, we defensively
  flush the block device before journal replay.

Tested: fxfs-large-truncate-tests (which is now enabled).
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588362
Original-Revision: 47af883f2f68d6358d8429bf58c1eb1daa224bc7
GitOrigin-RevId: 0b937721a4c7bd6e5c154091ae77d2f5a1e0f2da
Change-Id: I3310f8f8e1589bf42203e79e7af0cef949eee16b