[roll] Roll fuchsia [kernel][vm] Less specific oldest queue
The page queues are presently designed to pre-emptively process the
LRU queue in anticipation of aging needing a free slot. This means
that it cannot be assumed that pages will remain in the oldest queue.

Presently this issue cannot actually occur since this test will never
trigger the aging thread to wake up, and as a result the eviction
thread will never get triggered. However this is quite brittle and if
either of those assumptions change this test will become flaky.

Original-Bug: 85830
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589784
Original-Revision: b61b52bd686fa6789345c9f4855a823b07802f47
GitOrigin-RevId: 2800beb1b2efad5cabe330d2a19b95f85832dbcc
Change-Id: I015303324358e07d6eb87aaa32726602ab6f2cfb