[roll] Roll fuchsia [third_party][zstd] compile with ZSTD_MULTITHREAD when !is_fuchsia
We aim to enable multithreading in zstd when not compiling for fuchsia
(i.e., "host"); however, the current spelling of `is_host` fails to
cover the case of the linux toolchain. Recently, internal builds have
begun to build this library - in the context of the `zbi` tool - with
the linux toolchain and new errors have surfaced as a result. To better
reflect the original intent, `is_host` is replaced with  `!is_fuchsia`.
This change only mitigates the bug causing the error in the `zbi` tool;
a separate change will properly fix things there.

Original-Bug: b/202053801
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590383
Original-Revision: 1e06a3b4aa58ae191634857b1a8966bf7090e2b4
GitOrigin-RevId: 192feb605027ab0274a325b27dc8f07b24939e48
Change-Id: Icabb911196115647274545452f7e4645ca871611