[roll] Roll fuchsia [audio_core][test] Specify device_id and clock mode
Extend HermeticFidelityTest for multiple device configs and client clock
modes. Callers can specify the device ID to be targeted, as well as the
type of reference clock to be used with the renderer.

Updates to HermeticFidelityTest will be leveraged in an upcoming CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588224
Original-Revision: f76b5803a90bffcae3a293a51d9f5583cd50437a
GitOrigin-RevId: 76a102fd390cb776358cd1be528ce6ec65620206
Change-Id: Ic7eb3a001d2b6c5ffae52607d65818df974745aa