[roll] Roll fuchsia [zbi][multithread] Check upper bound for workers
Checks the upper bound for the number of workers before setting it.
Zbi tool will fail when built without the ZSTD_MULTITHREAD flag. This
occurs because the max number of workers is 0 in single threaded
compression and the number of workers was set to hardware_concurrency()

Original-Bug: b/202053801
Test: Built and ran ZBI tool without ZSTD_MULTITHREAD
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590387
Original-Revision: 755168084293f2b364f39d7b5b07f5c4415427f5
GitOrigin-RevId: 59bf473932cb0a87cc54773c4e5a2177005838c4
Change-Id: Ie75445e1be407741ad276e1dbb17abbc11e0f50c