[roll] Roll fuchsia [serial] Stop passing file handle to serial server
The serial server was throwing errors because botanist was closing the
serial log file before it was done copying to it. This CL fixes that
issue by having the server open and maintain its own file pointer,
instead of sharing the handle with botanist.

Original-Bug: 86049
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590391
Original-Revision: 9bcac99728224c8acc0cc661bf81acbfa5599550
GitOrigin-RevId: 46263a87de0e787ef5711e71a234b2f0a11d027f
Change-Id: I7aa5bf97dd7b63ff55e330aba07133b5de2a3cf2